Added the possibility of passing as input the eigenvectors that the define the PC hyperplane for each task
The function can take a matrix with the specific eigenvectors that
define the hyperplane for each task. The reason is that in some tasks
they may not appear in the same order, and therefore hide similarities.
For example, this seems to happen when comparing 1D to 2D lab tasks.
Tested (preliminarily) and working